Use bound for unprecise, but reasonable coordinate
Current method for get active window's coordinate does not get people
seeing window, just get the window in system hierarchy. So generally,
it does not see the menu bar and it cause coordinate error.

To solve this problem, adapt bound for coordinate guessing.
This is just work-around. Should alternate solution with right
solution in future.

But, at least, it works fine now.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>